storyboard-ui/src/main/webapp/app/entities/asset/detail/asset-detail.component.ts
import { Component, OnInit } from '@angular/core';
import { ActivatedRoute } from '@angular/router';

import { IAsset } from '../asset.model';
import { DataUtils } from 'app/core/util/data-util.service';

@Component({
  selector: 'jhi-asset-detail',
  templateUrl: './asset-detail.component.html',
})
export class AssetDetailComponent implements OnInit {
  asset: IAsset | null = null;

  constructor(protected dataUtils: DataUtils, protected activatedRoute: ActivatedRoute) {}

  ngOnInit(): void {
    this.activatedRoute.data.subscribe(({ asset }) => {
      this.asset = asset;
    });
  }

  byteSize(base64String: string): string {
    return this.dataUtils.byteSize(base64String);
  }

  openFile(base64String: string, contentType: string | null | undefined): void {
    this.dataUtils.openFile(base64String, contentType);
  }

  previousState(): void {
    window.history.back();
  }
}
